[#930] Rename DisplayCAL/setup.py to DisplayCAL/_setup.py
Sharing a basename with the root setup.py obscured that they play very
different roles (thin PEP 517 entry point vs. the actual distutils-based
data_files/install/py2app implementation) and produced an awkward
setup.setup() stutter at every import site. The leading underscore
mirrors the existing native_build.py/_native_build convention: the root
file is the public entry point, its underscore-prefixed counterpart is
internal implementation not meant to be invoked directly.
